+++ scripts/macros/monsterFeatures/generic/search.js
@@ -1,0 +1,3 @@
+export async function search({speaker, actor, token, character, item, args}) {
+    await this.actor.rollSkill('prc');
+}