+++ scripts/macros/monsterFeatures/undeadSpirit/undeadSpirit.js
@@ -1,0 +1,6 @@
+import {festeringAura} from './festeringAura.js';
+import {rottingClaw} from './rottingClaw.js';
+export let undeadSpirit = {
+    'festeringAura': festeringAura,
+    'rottingClaw': rottingClaw
+}